requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via pip-tools
dj-database-url==0.5.0
django-extensions==2.0.7
django==2.0.7
factory-boy==2.11.1
faker==0.8.17             # via factory-boy
first==2.0.1              # via pip-tools
flake8==3.5.0
mccabe==0.6.1             # via flake8
pip-tools==2.0.2
project-runpy==0.3.1
psycopg2-binary==2.7.5
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0           # via flake8
python-dateutil==2.7.3    # via faker
pytz==2018.5              # via django
six==1.11.0               # via django-extensions, faker, pip-tools, python-dateutil
text-unidecode==1.2       # via faker
